Filesystem read and write implementation.
Correctly determining the sector to write to from a disk address. The
function uses a ceil and divide to find the offset within a sector
where the IO operation starts.

Finished the implementation of the read and write data functions. They
use too many mem copy operations so it will be slow and inefficient for
now. It's using PIO mode anyway so it was always going to be
inefficient.

+++ filesystem.h
@@ -24,7 +24,7 @@ void write_data(uint8_t* data, uint64_t length, uint64_t startAddress);
 
 // Private methods
 
-uint64_t get_sector_from_address(uint64_t address, uint8_t* sectorOffset);
+uint64_t get_sector_from_address(uint64_t address, uint16_t* sectorOffset);
 
 // TODO : I/O api using the FILE* c std
